moved external files out of package
