+++ data/area21/exit.png
