--- quiz_server/templates/data.jinja
@@ -1,9 +1,0 @@
-{% extends "base.jinja" %}
-
-{% block content %}
-
-<main style="overflow-x: scroll; align-items: normal;">
-    {{ data_html }}
-</main>
-
-{% endblock %}
