+++ quiz_server/templates/list.jinja
@@ -5,13 +5,12 @@
     <ul>
         {% for quiz in quiz_list %}
         <li>
-            <a href="{{ url_for('quiz.get_quiz', quiz_name=quiz) }}">{{ quiz }}</a>
+            <a href="{{ url_for('quiz.quiz', quiz_name=quiz) }}">{{ quiz }}</a>
         </li>
         {% endfor %}
     </ul>
 
-    <a href="{{ url_for('quiz.get_data') }}">Get Data</a>
-    <a href="{{ url_for('quiz.get_results') }}">Get Results</a>
+    <a href="{{ url_for('quiz.info') }}">Info</a>
 </main>
 
 {% endblock %}
