+++ quiz_server/templates/quiz.jinja
@@ -58,6 +58,7 @@
     let QUIZ_NAME = "{{ quiz_name }}";
     let QUIZ_DATA = {{ quiz_data | tojson }};
     let STATIC_URL = "{{ url_for('static', filename='') }}";
+    let DATA_URL = "{{ url_for('quiz.data', filename='') }}";
 
     const ELEMENTS = {
         body: document.body,
@@ -110,7 +111,7 @@
     function updateIcon() {
         QUESTION_ICON = QUIZ_DATA[CURRENT_QUESTION - 1].logo_image;
         if (QUESTION_ICON) {
-            ELEMENTS.headerIcon.src = STATIC_URL + "images/" + QUESTION_ICON;
+            ELEMENTS.headerIcon.src = DATA_URL + QUESTION_ICON
             ELEMENTS.headerIcon.alt = QUIZ_NAME;
         }
         else {
@@ -122,7 +123,7 @@
     function updateBackground() {
         QUESTION_BACKGROUND = QUIZ_DATA[CURRENT_QUESTION - 1].bg_image;
         if (QUESTION_BACKGROUND) {
-            ELEMENTS.body.style.backgroundImage = `url('${STATIC_URL}images/${QUESTION_BACKGROUND}')`;
+            ELEMENTS.body.style.backgroundImage = `url('${DATA_URL}${QUESTION_BACKGROUND}')`;
         }
         else {
             ELEMENTS.body.style.backgroundImage = "none";
@@ -222,7 +223,7 @@
     function submitResults() {
         session_results = getResults();
 
-        fetch('{{ url_for('quiz.submit_session') }}', {
+        fetch('{{ url_for('quiz.submit') }}', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
